Check if a column exists before dropping the column in migration tables

+++ database/migrations/2020_07_16_191150_drop_column_in_products_table.php
@@ -13,9 +13,11 @@ class DropColumnInProductsTable extends Migration
      */
     public function up()
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('pix');
-        });
+        if (Schema::hasColumn('products', 'pix')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->dropColumn('pix');
+            });
+        }
     }
 
     /**
@@ -25,8 +27,10 @@ class DropColumnInProductsTable extends Migration
      */
     public function down()
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('pix');
-        });
+        if (Schema::hasColumn('products', 'pix')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->dropColumn('pix');
+            });
+        }
     }
 }
