+++ database/migrations/2020_07_16_191150_drop_columns_in_users_table.php
@@ -13,9 +13,11 @@ class DropColumnsInUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('tax_status');
-        });
+        if (Schema::hasColumn('users', 'tax_status')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('tax_status');
+            });
+        }
     }
 
     /**
@@ -25,8 +27,10 @@ class DropColumnsInUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('tax_status');
-        });
+        if (Schema::hasColumn('users', 'tax_status')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('tax_status');
+            });
+        }
     }
 }
